// backend/src/api/client/ApiTypes.ts
export type StrengthSlug =
  | 'carefulness'
  | 'compassion'
  | 'courage'
  | 'creativity'
  | 'curiosity'
  | 'enthusiasm'
  | 'fairness'
  | 'forgiveness'
  | 'gratitude'
  | 'grit'
  | 'honesty'
  | 'hope'
  | 'humour'
  | 'judgement'
  | 'kindness'
  | 'leadership'
  | 'love'
  | 'loveOfBeauty'
  | 'loveOfLearning'
  | 'modesty'
  | 'perseverance'
  | 'perspective'
  | 'selfRegulation'
  | 'socialIntelligence'
  | 'spirituality'
  | 'teamwork';
// ...
export type ArticleCategory = {
  id: string;
  subCategories: ArticleCategory[];
  thumbnail: string;
  displayAs: 'list' | 'grid';
  translations: ArticleCategoryTranslation[];
  order: number;
  articles: Array<{
    id: string;
    translations: Array<{
      language: LanguageCode;
      title: string;
      description: string;
      thumbnail?: string;
    }>;
    order: number;
    thumbnail: string;
    length: string;
    strengths: StrengthSlug[];
    isHidden: boolean;
    isLocked: boolean;
  }>;
  categoryPath: Array<{
    id: string;
    translations: Array<{
      language: LanguageCode;
      name: string;
    }>;
  }>;
  isHidden: boolean;
  isLocked: boolean;
};
export type ArticleCategoryTranslation = {
  language: LanguageCode;
  name: string;
  description: string;
  thumbnail?: string;
};
// ...
export type LanguageCode = 'fi' | 'en' | 'sv';
// ...
export type GetArticleCategoryResponse = ArticleCategory;